COMPIL.MAC for BCPL and Mon 7.04
COMPIL.MAC is a modified source for the COMPIL command on TOPS-10 Monitor 7.04 so that it will automatically compile BCPL source from the monitor prompt. You will need to compile this source to produce a new COMPIL.EXE, which you can the put in the SYS: volume. Once this is done COMPIL will support BCPL, allowing you to simple EXECUTE SOURCE.BCPL/BCPL and it will automatically compile, link and execute your BCPL source.